Rip out v1 voice room system end-to-end
Drops the voice_rooms table, get_available_voice_rooms RPC, the v1 branch
of /api/voice/token, the v1 voice hooks/services (useAvailableVoiceRooms,
useLoungeRoomId, the {roomId} shape of useVoiceToken), the
AvailableVoiceRoom and VoiceRoom* type aliases, the lounge concept, and
the /admin/voice/[id] + /gedu/voice/[id] routes. /api/voice/token is now
{groupId}-only; the v1 mid-session enrollment gate (which fired
"Your enrollment starts next session" for users joining a v2 club whose
session was already in progress) is gone — v2 treats active membership
as the binary access predicate. VoiceSessionPage drops the tokenMode
prop and the JS auto-leave loop; Daily's token exp is the hard cutoff.

Migration 00060 drops voice_rooms (CASCADE handles its RLS/indexes/
publication entry), drops get_available_voice_rooms, redefines
get_my_groups without voice_room_id, and patches commit_group_changes
to stop INSERTing into the now-gone table.

The admin/gedu/customer v1 groups UI keeps its Join buttons + lounge
cards visible but neutered (disabled, no-op). A TODO.md item tracks
tearing out that whole surface once it's confirmed unused.
