+++ TODO.md
@@ -2,6 +2,7 @@
 
 ## Cleanup
 
+- [ ] **Tear out the v1 groups UI now that its voice surface is a no-op.** When the v1 voice room system (table + RPCs + token branch) was deleted, the Join buttons and lounge cards on the admin/gedu/customer v1 groups surface were left in place but neutered (rendered disabled, no click handler). The pages are still navigable and still display group/lounge metadata via `useMyGroups` + `useGroupsWithVoice`, but a user clicking "Join" gets nothing. Cleanup: delete `src/app/(dashboard)/admin/groups/`, `src/app/(dashboard)/gedu/groups/`, `src/app/(dashboard)/parent/gamers/page.tsx` and `src/app/(dashboard)/parent/groups/[id]/page.tsx` along with their content components (`AdminGroupsPageContent`, `GeduGroupsPageContent`, `AdminGroupDetailContent`, `GeduGroupDetailContent`, `CustomerGroupDetailContent`), the shared `GroupsListContent`/`GroupDetailContent`/`LoungeCard`/`GroupCard` UI primitives, the `useGroupsWithVoice` hook, and the `useMyGroups` query — plus their tests and translation keys (`groups.adminLounge*`, `groups.geduLounge*`, `groups.allGroups*`, `groups.yourGroups*`). Also drop `voiceIsOpen`/`voiceNextSessionStart` plumbing from `GroupCard` and remove the `lounges` prop from `GroupsListContent` (currently empty arrays from both call sites). Coordinates with the existing "Delete orphaned parent/gamer sub-routes" item below — these are the same routes.
 - [ ] **`GroupDetailContent.tsx:190` hides age display when gamer has no gender.** With migration `00055` (gender nullable) and the Add Gamer v1 flow making gender optional, the `gamer.dateOfBirth && gamer.gender` check now silently drops the entire age/gender detail row for v1 gamers — both pieces of info disappear, not just gender. Fix is straightforward (split into independent conditional spans, render the wrapper if either is present) but the file is slated for deprecation, so noted here rather than patched. While doing it, also retype `GroupGamer.gender` in `groups.service.ts:10` from `string` to `string | null` so future consumers get a compile-time signal about the null case.
 - [ ] **Delete `SwitchToGamerDialog` once the family selector fully replaces it.** `src/components/customer/SwitchToGamerDialog.tsx` is the older confirmation-modal path for parent→gamer switching, used by `parent/gamers/page.tsx` and `CustomerGroupDetailContent.tsx`. The new `FamilyProfileSelector` (on `/parent#my-family` and `/gamer#my-family`) does the same switch in one click against the same `/api/auth/switch-account` endpoint. Two paths to the same intent risks divergence (e.g. if a parent-PIN gate is added for gamer→parent, it's easy to add it in one path and forget the other). Kill the dialog after deciding what should replace its in-context entry points (or just inline a one-tile selector in those spots), and prune the related `theme-audit-findings.md` follow-up about promoting its `info`-coloured button to a Button variant.
 - [ ] **Delete orphaned parent/gamer sub-routes left behind by the unified-navbar single-page-scroll redesign.** When the parent and gamer dashboards moved to the in-page section-pill UX (My Family / [Yty] / Feedback / Settings on `/parent` and `/gamer`), the sidebar and the dashboard quick-action cards that linked to these routes were removed. The routes themselves still resolve, but no in-app navigation points at them anymore — they're reachable only by typing the URL or following an external/email link. Replace each one (or its functionality) inside the new dashboard sections, then delete:
