+++ src/app/(dashboard)/gamer/voice/[id]/page.tsx
@@ -9,13 +9,13 @@ export async function generateMetadata(): Promise<Metadata> {
 }
 
 /**
- * Gamer voice room. The `[id]` segment is `product_groups_v2.id` — the v2
- * token endpoint derives the Daily room name from the group + the current
- * session window, so there's no `voice_rooms` row to look up.
+ * Gamer voice room. The `[id]` segment is `product_groups_v2.id` — the
+ * token endpoint derives the Daily room name from the group + current
+ * session window, so there is no DB-side voice room row to look up.
  *
- * Back link goes to the gamer dashboard (`/gamer`) — the Sessions section
- * is the only in-app surface that produces voice links now that the
- * `/gamer/groups` deprecation is underway (TODO.md "Cleanup" item).
+ * Back link goes to the gamer dashboard — the Sessions section is the
+ * only in-app surface that produces voice links now (`/gamer/groups`
+ * deprecation is underway; see TODO.md).
  */
 export default async function GamerVoiceSessionPage({
   params,
@@ -23,11 +23,5 @@ export default async function GamerVoiceSessionPage({
   params: Promise<{ id: string }>;
 }) {
   const { id } = await params;
-  return (
-    <VoiceSessionPage
-      roomId={id}
-      backHref={ROUTES.gamer.dashboard}
-      tokenMode="v2"
-    />
-  );
+  return <VoiceSessionPage groupId={id} backHref={ROUTES.gamer.dashboard} />;
 }
