+++ src/components/gedu/GeduGroupDetailContent.tsx
@@ -19,7 +19,6 @@ export function GeduGroupDetailContent({ groupId }: GeduGroupDetailContentProps)
       isLoading={isLoading}
       error={error}
       backHref={ROUTES.gedu.groups}
-      voiceRoute={ROUTES.gedu.voiceSession}
     />
   );
 }
