+++ src/services/voice/index.ts
@@ -1,9 +1,2 @@
 export { VoiceService } from "./voice.service";
-export type { AvailableVoiceRoomWithWindow } from "./voice.service";
-export {
-  voiceKeys,
-  useAvailableVoiceRooms,
-  useVoiceToken,
-  useVoiceTokenV2,
-  useLoungeRoomId,
-} from "./voice.queries";
+export { voiceKeys, useVoiceToken } from "./voice.queries";
