+++ tests/integration/api/voice-token.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import { POST } from "@/app/api/voice/token/route";
 import { NextResponse } from "next/server";
+import { POST } from "@/app/api/voice/token/route";
+import { DailyApiError } from "@/lib/daily";
 import { mockSupabaseSuccess } from "../../mocks/supabase";
-import { createMockVoiceRoom } from "../../mocks/voice";
 import { VOICE_CONFIG } from "@/lib/constants/voice";
 
 // --- Mocks ---
@@ -13,40 +13,40 @@ vi.mock("@/lib/auth", () => ({
 }));
 
 const mockAdminFrom = vi.fn();
-const mockAdminRpc = vi.fn();
 vi.mock("@/lib/supabase/admin", () => ({
   createAdminClient: vi.fn(() => ({
     from: (...args: unknown[]) => mockAdminFrom(...args),
-    rpc: (...args: unknown[]) => mockAdminRpc(...args),
   })),
 }));
 
 const mockCreateMeetingToken = vi.fn();
-const mockGetDailyRoom = vi.fn();
 const mockCreateDailyRoom = vi.fn();
-// Stub the network-touching helpers but keep `buildUserName` real so any
-// future change to its sanitization is exercised by the existing userName
-// assertions instead of silently passing against a stale mirror.
 vi.mock("@/lib/daily", async () => {
   const actual = await vi.importActual<typeof import("@/lib/daily")>("@/lib/daily");
   return {
     ...actual,
     createMeetingToken: (...args: unknown[]) => mockCreateMeetingToken(...args),
-    getDailyRoom: (...args: unknown[]) => mockGetDailyRoom(...args),
     createDailyRoom: (...args: unknown[]) => mockCreateDailyRoom(...args),
   };
 });
 
 const mockComputeSessionWindow = vi.fn();
-const mockIsEnrolledForSession = vi.fn();
-vi.mock("@/lib/session-schedule", () => ({
-  computeSessionWindow: (...args: unknown[]) => mockComputeSessionWindow(...args),
-  isEnrolledForSession: (...args: unknown[]) => mockIsEnrolledForSession(...args),
-}));
+vi.mock("@/lib/session-schedule", async () => {
+  const actual = await vi.importActual<typeof import("@/lib/session-schedule")>(
+    "@/lib/session-schedule",
+  );
+  return {
+    ...actual,
+    computeSessionWindow: (...args: unknown[]) => mockComputeSessionWindow(...args),
+  };
+});
 
 // --- Helpers ---
 
-function createTokenRequest(body: Record<string, unknown>): Request {
+const GROUP_ID = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee";
+const PRODUCT_ID = "11111111-2222-3333-4444-555555555555";
+
+function tokenRequest(body: Record<string, unknown>): Request {
   return new Request("http://localhost:3000/api/voice/token", {
     method: "POST",
     headers: { "Content-Type": "application/json" },
@@ -54,9 +54,9 @@ function createTokenRequest(body: Record<string, unknown>): Request {
   });
 }
 
-function mockAuthenticatedWithProfile(
+function authAs(
   userId: string,
-  profile: { role: string; first_name: string; username: string | null },
+  profile: { role: string; first_name: string },
 ) {
   if (profile.role === "customer") {
     mockRequireRole.mockResolvedValue(
@@ -67,7 +67,6 @@ function mockAuthenticatedWithProfile(
     );
     return;
   }
-
   mockRequireRole.mockResolvedValue({
     user: { id: userId },
     profile,
@@ -75,57 +74,79 @@ function mockAuthenticatedWithProfile(
   });
 }
 
-function mockRoomLookup(room: Record<string, unknown> | null) {
-  if (room) {
-    mockAdminFrom.mockReturnValue({
-      select: vi.fn().mockReturnValue({
-        eq: vi.fn().mockReturnValue({
-          single: vi.fn().mockResolvedValue(mockSupabaseSuccess(room)),
-        }),
-      }),
-    });
-  } else {
-    mockAdminFrom.mockReturnValue({
-      select: vi.fn().mockReturnValue({
-        eq: vi.fn().mockReturnValue({
-          single: vi.fn().mockResolvedValue({
-            data: null,
-            error: { message: "No rows", code: "PGRST116" },
+/**
+ * Build the `from(...)` router used by the route. Three tables in order:
+ * `product_groups_v2`, `participations_v2` (gamer-only), and
+ * `gedu_group_assignments_v2` (gedu-only) — the mock dispatches by table
+ * name and returns the matching thenable shape.
+ */
+function mockTables(opts: {
+  group: {
+    timezone?: string;
+    is_remote?: boolean;
+    slots?: Array<{ weekday: number; start_time: string; duration_minutes: number }>;
+  } | null;
+  participation?: { id: string } | null;
+  geduAssignment?: { group_id: string } | null;
+}) {
+  mockAdminFrom.mockImplementation((table: string) => {
+    if (table === "product_groups_v2") {
+      const row = opts.group
+        ? {
+            id: GROUP_ID,
+            product_id: PRODUCT_ID,
+            product: {
+              id: PRODUCT_ID,
+              timezone: opts.group.timezone ?? "Europe/Helsinki",
+              is_remote: opts.group.is_remote ?? true,
+              slots: opts.group.slots ?? [
+                { weekday: 1, start_time: "14:00", duration_minutes: 60 },
+              ],
+            },
+          }
+        : null;
+      return {
+        select: vi.fn().mockReturnValue({
+          eq: vi.fn().mockReturnValue({
+            maybeSingle: vi.fn().mockResolvedValue(mockSupabaseSuccess(row)),
           }),
         }),
-      }),
-    });
-  }
-}
-
-function createGroupRoom(overrides: Record<string, unknown> = {}) {
-  return {
-    ...createMockVoiceRoom(),
-    room_type: "group",
-    group_id: "group-uuid-1234",
-    product_groups: {
-      gedu_id: "gedu-user-id",
-      product_id: "product-uuid-1234",
-      products: {
-        day_of_week: 1,
-        start_time: "14:00",
-        timezone: "Europe/Helsinki",
-        duration_minutes: 60,
-      },
-    },
-    ...overrides,
-  };
-}
-
-function createSpecialRoom(roomType: string) {
-  return {
-    ...createMockVoiceRoom({
-      group_id: null,
-      room_type: roomType,
-      daily_room_name: roomType === "admin_only" ? "admin-lounge" : "gedu-lounge",
-    }),
-    product_groups: null,
-  };
+      };
+    }
+    if (table === "participations_v2") {
+      return {
+        select: vi.fn().mockReturnValue({
+          eq: vi.fn().mockReturnValue({
+            eq: vi.fn().mockReturnValue({
+              eq: vi.fn().mockReturnValue({
+                limit: vi.fn().mockReturnValue({
+                  maybeSingle: vi
+                    .fn()
+                    .mockResolvedValue(mockSupabaseSuccess(opts.participation ?? null)),
+                }),
+              }),
+            }),
+          }),
+        }),
+      };
+    }
+    if (table === "gedu_group_assignments_v2") {
+      return {
+        select: vi.fn().mockReturnValue({
+          eq: vi.fn().mockReturnValue({
+            eq: vi.fn().mockReturnValue({
+              limit: vi.fn().mockReturnValue({
+                maybeSingle: vi
+                  .fn()
+                  .mockResolvedValue(mockSupabaseSuccess(opts.geduAssignment ?? null)),
+              }),
+            }),
+          }),
+        }),
+      };
+    }
+    return {};
+  });
 }
 
 // --- Tests ---
@@ -137,643 +158,231 @@ describe("POST /api/voice/token", () => {
     vi.clearAllMocks();
     process.env.NEXT_PUBLIC_DAILY_DOMAIN = "testdomain";
     mockCreateMeetingToken.mockResolvedValue("mock-daily-token");
-    mockGetDailyRoom.mockResolvedValue({ name: "test-room" });
+    mockCreateDailyRoom.mockResolvedValue({ name: "test-room" });
+    // Default window: open right now, closes in an hour.
+    const nextStart = new Date(Date.now() - 60_000);
     mockComputeSessionWindow.mockReturnValue({
       isOpen: true,
-      nextSessionStart: new Date(),
-      windowOpensAt: new Date(),
-      windowClosesAt: new Date(Date.now() + 3600_000),
+      nextSessionStart: nextStart,
+      windowOpensAt: new Date(nextStart.getTime() - 300_000),
+      windowClosesAt: new Date(nextStart.getTime() + 3600_000),
     });
-    mockIsEnrolledForSession.mockReturnValue(true);
   });
 
   afterEach(() => {
     process.env.NEXT_PUBLIC_DAILY_DOMAIN = originalEnv;
   });
 
-  it("should return 401 when not authenticated", async () => {
+  it("returns 401 when not authenticated", async () => {
     mockRequireRole.mockResolvedValue(
       NextResponse.json({ error: "Unauthorized" }, { status: 401 }),
     );
-
-    const response = await POST(createTokenRequest({ roomId: "room-1" }));
-    const data = await response.json();
-
-    expect(response.status).toBe(401);
-    expect(data.error).toBe("Unauthorized");
+    const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+    expect(res.status).toBe(401);
   });
 
-  it("should return 403 for customer role", async () => {
-    mockAuthenticatedWithProfile("customer-id", {
-      role: "customer",
-      first_name: "Parent",
-      username: "parent1",
-    });
-
-    const response = await POST(createTokenRequest({ roomId: "room-1" }));
-    const data = await response.json();
-
-    expect(response.status).toBe(403);
-    expect(data.error).toBe("You do not have permission to join voice rooms");
+  it("returns 403 for customer role", async () => {
+    authAs("customer-id", { role: "customer", first_name: "Parent" });
+    const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+    expect(res.status).toBe(403);
   });
 
-  it("should return 400 when roomId is missing", async () => {
-    mockAuthenticatedWithProfile("gedu-user-id", {
-      role: "gedu",
-      first_name: "Educator",
-      username: "edu1",
-    });
-
-    const response = await POST(createTokenRequest({}));
-    const data = await response.json();
-
-    expect(response.status).toBe(400);
-    expect(data.error).toBe("roomId is required");
+  it("returns 400 when groupId is missing", async () => {
+    authAs("gedu-id", { role: "gedu", first_name: "Edu" });
+    const res = await POST(tokenRequest({}));
+    expect(res.status).toBe(400);
   });
 
-  it("should return 404 when room does not exist", async () => {
-    mockAuthenticatedWithProfile("gedu-user-id", {
-      role: "gedu",
-      first_name: "Educator",
-      username: "edu1",
-    });
-    mockRoomLookup(null);
-
-    const response = await POST(createTokenRequest({ roomId: "nonexistent" }));
-    const data = await response.json();
-
-    expect(response.status).toBe(404);
+  it("returns 404 when the group does not exist", async () => {
+    authAs("gedu-id", { role: "gedu", first_name: "Edu" });
+    mockTables({ group: null });
+    const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+    const data = await res.json();
+    expect(res.status).toBe(404);
     expect(data.error).toBe("Room not found");
   });
 
-  it("should return 500 when NEXT_PUBLIC_DAILY_DOMAIN is not configured", async () => {
-    delete process.env.NEXT_PUBLIC_DAILY_DOMAIN;
-
-    mockAuthenticatedWithProfile("gedu-user-id", {
-      role: "gedu",
-      first_name: "Educator",
-      username: "edu1",
-    });
-    mockRoomLookup(createGroupRoom());
-
-    const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-    const data = await response.json();
-
-    expect(response.status).toBe(500);
-    expect(data.error).toBe("Voice chat is not configured");
+  it("returns 404 when the product is in-person (is_remote = false)", async () => {
+    authAs("gedu-id", { role: "gedu", first_name: "Edu" });
+    mockTables({ group: { is_remote: false }, geduAssignment: { group_id: GROUP_ID } });
+    const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+    expect(res.status).toBe(404);
   });
 
-  describe("admin_only room", () => {
-    it("should allow admin to join", async () => {
-      mockAuthenticatedWithProfile("admin-user-id", {
-        role: "admin",
-        first_name: "Admin",
-        username: "admin1",
-      });
-      mockRoomLookup(createSpecialRoom("admin_only"));
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      expect(response.status).toBe(200);
-    });
-
-    it("should reject gedu from admin-only room", async () => {
-      mockAuthenticatedWithProfile("gedu-user-id", {
-        role: "gedu",
-        first_name: "Educator",
-        username: "edu1",
-      });
-      mockRoomLookup(createSpecialRoom("admin_only"));
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      const data = await response.json();
-
-      expect(response.status).toBe(403);
-      expect(data.error).toBe("Only admins can join this room");
-    });
-
-    it("should reject gamer from admin-only room", async () => {
-      mockAuthenticatedWithProfile("gamer-user-id", {
-        role: "gamer",
-        first_name: "Gamer",
-        username: "gamer1",
-      });
-      mockRoomLookup(createSpecialRoom("admin_only"));
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      expect(response.status).toBe(403);
-    });
-  });
-
-  describe("gedu_only room", () => {
-    it("should allow gedu to join", async () => {
-      mockAuthenticatedWithProfile("gedu-user-id", {
-        role: "gedu",
-        first_name: "Educator",
-        username: "edu1",
-      });
-      mockRoomLookup(createSpecialRoom("gedu_only"));
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      expect(response.status).toBe(200);
-    });
-
-    it("should allow admin to join gedu-only room", async () => {
-      mockAuthenticatedWithProfile("admin-user-id", {
-        role: "admin",
-        first_name: "Admin",
-        username: "admin1",
-      });
-      mockRoomLookup(createSpecialRoom("gedu_only"));
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      expect(response.status).toBe(200);
-    });
-
-    it("should reject gamer from gedu-only room", async () => {
-      mockAuthenticatedWithProfile("gamer-user-id", {
-        role: "gamer",
-        first_name: "Gamer",
-        username: "gamer1",
-      });
-      mockRoomLookup(createSpecialRoom("gedu_only"));
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      const data = await response.json();
-
-      expect(response.status).toBe(403);
-      expect(data.error).toBe("Only educators and admins can join this room");
-    });
-  });
-
-  describe("group room — gedu access", () => {
-    it("should allow assigned gedu to join", async () => {
-      mockAuthenticatedWithProfile("gedu-user-id", {
-        role: "gedu",
-        first_name: "Educator",
-        username: "edu1",
-      });
-      mockRoomLookup(createGroupRoom());
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      expect(response.status).toBe(200);
-    });
-
-    it("should reject unassigned gedu", async () => {
-      mockAuthenticatedWithProfile("other-gedu-id", {
-        role: "gedu",
-        first_name: "Other Educator",
-        username: "edu2",
-      });
-      mockRoomLookup(createGroupRoom());
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      const data = await response.json();
-
-      expect(response.status).toBe(403);
-      expect(data.error).toBe("You are not assigned to this group");
-    });
-  });
-
-  describe("group room — gamer access", () => {
-    /** Helper: mock both room lookup + enrollment query for gamer tests */
-    function mockGamerEnrollment(enrollment: { id: string; created_at: string } | null) {
-      mockAdminFrom.mockImplementation((table: string) => {
-        if (table === "voice_rooms") {
-          return {
-            select: vi.fn().mockReturnValue({
-              eq: vi.fn().mockReturnValue({
-                single: vi.fn().mockResolvedValue(mockSupabaseSuccess(createGroupRoom())),
-              }),
-            }),
-          };
-        }
-        if (table === "group_enrollments") {
-          return {
-            select: vi.fn().mockReturnValue({
-              eq: vi.fn().mockReturnValue({
-                eq: vi.fn().mockReturnValue({
-                  eq: vi.fn().mockReturnValue({
-                    limit: vi.fn().mockReturnValue({
-                      maybeSingle: vi.fn().mockResolvedValue(mockSupabaseSuccess(enrollment)),
-                    }),
-                  }),
-                }),
-              }),
-            }),
-          };
-        }
-        return {};
-      });
-    }
-
-    it("should allow enrolled gamer to join when session is open", async () => {
-      mockAuthenticatedWithProfile("gamer-user-id", {
-        role: "gamer",
-        first_name: "Gamer",
-        username: "gamer1",
-      });
-      mockRoomLookup(createGroupRoom());
-
-      // Enrolled well before session start
-      const sessionStart = new Date();
-      mockComputeSessionWindow.mockReturnValue({
-        isOpen: true,
-        nextSessionStart: sessionStart,
-        windowOpensAt: new Date(sessionStart.getTime() - 300_000),
-        windowClosesAt: new Date(sessionStart.getTime() + 3600_000),
-      });
-
-      mockGamerEnrollment({
-        id: "enrollment-1",
-        created_at: new Date(sessionStart.getTime() - 7 * 24 * 3600_000).toISOString(),
-      });
-      mockIsEnrolledForSession.mockReturnValue(true);
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      expect(response.status).toBe(200);
-    });
-
-    it("should reject unenrolled gamer", async () => {
-      mockAuthenticatedWithProfile("gamer-user-id", {
-        role: "gamer",
-        first_name: "Gamer",
-        username: "gamer1",
-      });
-
-      mockGamerEnrollment(null);
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      const data = await response.json();
-
-      expect(response.status).toBe(403);
+  describe("membership gate", () => {
+    it("rejects a gamer with no active participation", async () => {
+      authAs("gamer-id", { role: "gamer", first_name: "Kid" });
+      mockTables({ group: {}, participation: null });
+      const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+      const data = await res.json();
+      expect(res.status).toBe(403);
       expect(data.error).toBe("You are not enrolled in this group");
     });
 
-    it("should reject gamer outside session window", async () => {
-      mockAuthenticatedWithProfile("gamer-user-id", {
-        role: "gamer",
-        first_name: "Gamer",
-        username: "gamer1",
-      });
+    it("rejects a gedu not assigned to the product", async () => {
+      authAs("gedu-id", { role: "gedu", first_name: "Edu" });
+      mockTables({ group: {}, geduAssignment: null });
+      const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+      const data = await res.json();
+      expect(res.status).toBe(403);
+      expect(data.error).toBe("You are not assigned to this group");
+    });
 
+    it("admin bypasses the membership check", async () => {
+      authAs("admin-id", { role: "admin", first_name: "Boss" });
+      mockTables({ group: {} });
+      const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+      expect(res.status).toBe(200);
+    });
+
+    it("allows an active participant regardless of how recently they signed up", async () => {
+      // v2 dropped the v1 mid-session enrollment gate — active membership
+      // is the binary access predicate, so a gamer who joined 30s ago
+      // gets in just like one who joined a week ago.
+      authAs("gamer-id", { role: "gamer", first_name: "Kid" });
+      mockTables({
+        group: {},
+        participation: { id: "participation-1" },
+      });
+      const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+      expect(res.status).toBe(200);
+    });
+  });
+
+  describe("session window gate", () => {
+    it("rejects when no slot has an open window", async () => {
+      authAs("gedu-id", { role: "gedu", first_name: "Edu" });
+      mockTables({ group: {}, geduAssignment: { group_id: GROUP_ID } });
       mockComputeSessionWindow.mockReturnValue({
         isOpen: false,
         nextSessionStart: new Date(Date.now() + 86400_000),
         windowOpensAt: new Date(Date.now() + 86100_000),
         windowClosesAt: new Date(Date.now() + 90000_000),
       });
-
-      mockGamerEnrollment({
-        id: "enrollment-1",
-        created_at: new Date(Date.now() - 7 * 24 * 3600_000).toISOString(),
-      });
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      const data = await response.json();
-
-      expect(response.status).toBe(403);
+      const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+      const data = await res.json();
+      expect(res.status).toBe(403);
       expect(data.error).toBe("Room is not open yet");
     });
 
-    it("should reject gamer who enrolled after session started (mid-session enrollment)", async () => {
-      mockAuthenticatedWithProfile("gamer-user-id", {
-        role: "gamer",
-        first_name: "Gamer",
-        username: "gamer1",
+    it("picks the first slot whose window is open when multiple slots exist", async () => {
+      authAs("gedu-id", { role: "gedu", first_name: "Edu" });
+      mockTables({
+        group: {
+          slots: [
+            { weekday: 1, start_time: "23:00", duration_minutes: 60 }, // Mon 11pm
+            { weekday: 2, start_time: "05:00", duration_minutes: 60 }, // Tue 5am
+          ],
+        },
+        geduAssignment: { group_id: GROUP_ID },
       });
-
-      // Session started 20 minutes ago
-      const sessionStart = new Date(Date.now() - 20 * 60_000);
-      mockComputeSessionWindow.mockReturnValue({
+      const closedWindow = {
+        isOpen: false,
+        nextSessionStart: new Date(Date.now() + 86400_000),
+        windowOpensAt: new Date(Date.now() + 86100_000),
+        windowClosesAt: new Date(Date.now() + 90000_000),
+      };
+      const openWindow = {
         isOpen: true,
-        nextSessionStart: sessionStart,
-        windowOpensAt: new Date(sessionStart.getTime() - 300_000),
-        windowClosesAt: new Date(sessionStart.getTime() + 3600_000),
-      });
-
-      // Enrolled 5 minutes ago — after the session started
-      mockGamerEnrollment({
-        id: "enrollment-1",
-        created_at: new Date(Date.now() - 5 * 60_000).toISOString(),
-      });
-      mockIsEnrolledForSession.mockReturnValue(false);
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      const data = await response.json();
-
-      expect(response.status).toBe(403);
-      expect(data.error).toBe("Your enrollment starts next session");
-    });
-
-    it("should allow gamer who enrolled just before session started", async () => {
-      mockAuthenticatedWithProfile("gamer-user-id", {
-        role: "gamer",
-        first_name: "Gamer",
-        username: "gamer1",
-      });
-
-      // Session started 1 minute ago
-      const sessionStart = new Date(Date.now() - 60_000);
-      mockComputeSessionWindow.mockReturnValue({
-        isOpen: true,
-        nextSessionStart: sessionStart,
-        windowOpensAt: new Date(sessionStart.getTime() - 300_000),
-        windowClosesAt: new Date(sessionStart.getTime() + 3600_000),
-      });
-
-      // Enrolled 2 minutes ago — before the session started
-      mockGamerEnrollment({
-        id: "enrollment-1",
-        created_at: new Date(Date.now() - 2 * 60_000).toISOString(),
-      });
-      mockIsEnrolledForSession.mockReturnValue(true);
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      expect(response.status).toBe(200);
+        nextSessionStart: new Date(Date.now() - 60_000),
+        windowOpensAt: new Date(Date.now() - 300_000),
+        windowClosesAt: new Date(Date.now() + 3600_000),
+      };
+      mockComputeSessionWindow
+        .mockReturnValueOnce(closedWindow)
+        .mockReturnValueOnce(openWindow);
+      const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+      expect(res.status).toBe(200);
     });
   });
 
-  describe("admin access", () => {
-    it("should allow admin to join any group room when session is open", async () => {
-      mockAuthenticatedWithProfile("admin-user-id", {
-        role: "admin",
-        first_name: "Admin User",
-        username: "admin1",
+  describe("happy path + Daily mechanics", () => {
+    it("lazy-creates the Daily room with exp = windowClosesAt + grace, then mints a gamer non-owner token", async () => {
+      authAs("gamer-id", { role: "gamer", first_name: "Kid" });
+      mockTables({
+        group: {},
+        participation: { id: "participation-1" },
       });
-      mockRoomLookup(createGroupRoom());
+      const windowClosesAt = new Date(Date.now() + 3600_000);
+      mockComputeSessionWindow.mockReturnValue({
+        isOpen: true,
+        nextSessionStart: new Date(Date.now() - 60_000),
+        windowOpensAt: new Date(Date.now() - 300_000),
+        windowClosesAt,
+      });
 
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      const data = await response.json();
+      const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+      const data = await res.json();
 
-      expect(response.status).toBe(200);
-      expect(data.role).toBe("admin");
+      expect(res.status).toBe(200);
+      expect(data.role).toBe("gamer");
+      expect(data.token).toBe("mock-daily-token");
+
+      const expectedExp =
+        Math.round(windowClosesAt.getTime() / 1000) +
+        VOICE_CONFIG.TOKEN_EXPIRY_GRACE_SECONDS;
+
+      expect(mockCreateDailyRoom).toHaveBeenCalledWith(
+        expect.objectContaining({
+          name: expect.stringMatching(/^g-aaaaaaaa-\d{12}$/),
+          expUnix: expectedExp,
+        }),
+      );
       expect(mockCreateMeetingToken).toHaveBeenCalledWith(
         expect.objectContaining({
-          isOwner: true,
-          userName: "admin-user-id|admin|Admin User",
+          isOwner: false,
+          expUnix: expectedExp,
+          userName: "gamer-id|gamer|Kid",
         }),
       );
     });
 
-    it("should reject admin outside session window", async () => {
-      mockAuthenticatedWithProfile("admin-user-id", {
-        role: "admin",
-        first_name: "Admin User",
-        username: "admin1",
+    it("mints an owner token for an assigned gedu", async () => {
+      authAs("gedu-id", { role: "gedu", first_name: "Edu" });
+      mockTables({
+        group: {},
+        geduAssignment: { group_id: GROUP_ID },
       });
-      mockRoomLookup(createGroupRoom());
-      mockComputeSessionWindow.mockReturnValue({
-        isOpen: false,
-        nextSessionStart: new Date(Date.now() + 86400_000),
-        windowOpensAt: new Date(Date.now() + 86100_000),
-        windowClosesAt: new Date(Date.now() + 90000_000),
-      });
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      const data = await response.json();
-
-      expect(response.status).toBe(403);
-      expect(data.error).toBe("Room is not open yet");
-    });
-  });
-
-  describe("gedu session window", () => {
-    it("should reject assigned gedu outside session window", async () => {
-      mockAuthenticatedWithProfile("gedu-user-id", {
-        role: "gedu",
-        first_name: "Educator",
-        username: "edu1",
-      });
-      mockRoomLookup(createGroupRoom());
-      mockComputeSessionWindow.mockReturnValue({
-        isOpen: false,
-        nextSessionStart: new Date(Date.now() + 86400_000),
-        windowOpensAt: new Date(Date.now() + 86100_000),
-        windowClosesAt: new Date(Date.now() + 90000_000),
-      });
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      const data = await response.json();
-
-      expect(response.status).toBe(403);
-      expect(data.error).toBe("Room is not open yet");
-    });
-  });
-
-  describe("token properties", () => {
-    it("should create owner token for gedu", async () => {
-      mockAuthenticatedWithProfile("gedu-user-id", {
-        role: "gedu",
-        first_name: "Test Educator",
-        username: "testgedu",
-      });
-      mockRoomLookup(createGroupRoom());
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-      const data = await response.json();
-
-      expect(response.status).toBe(200);
-      expect(data.token).toBe("mock-daily-token");
+      const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+      expect(res.status).toBe(200);
       expect(mockCreateMeetingToken).toHaveBeenCalledWith(
         expect.objectContaining({ isOwner: true }),
       );
     });
 
-    it("should create non-owner token for gamer", async () => {
-      mockAuthenticatedWithProfile("gamer-user-id", {
-        role: "gamer",
-        first_name: "Test Gamer",
-        username: "testgamer",
-      });
+    it("swallows a Daily 409 (room already exists) and still mints a token", async () => {
+      authAs("admin-id", { role: "admin", first_name: "Boss" });
+      mockTables({ group: {} });
+      mockCreateDailyRoom.mockRejectedValue(new DailyApiError(409, "Room exists"));
 
-      mockAdminFrom.mockImplementation((table: string) => {
-        if (table === "voice_rooms") {
-          return {
-            select: vi.fn().mockReturnValue({
-              eq: vi.fn().mockReturnValue({
-                single: vi.fn().mockResolvedValue(mockSupabaseSuccess(createGroupRoom())),
-              }),
-            }),
-          };
-        }
-        if (table === "group_enrollments") {
-          return {
-            select: vi.fn().mockReturnValue({
-              eq: vi.fn().mockReturnValue({
-                eq: vi.fn().mockReturnValue({
-                  eq: vi.fn().mockReturnValue({
-                    limit: vi.fn().mockReturnValue({
-                      maybeSingle: vi.fn().mockResolvedValue(mockSupabaseSuccess({
-                        id: "enrollment-1",
-                        created_at: new Date(Date.now() - 7 * 24 * 3600_000).toISOString(),
-                      })),
-                    }),
-                  }),
-                }),
-              }),
-            }),
-          };
-        }
-        return {};
-      });
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-
-      expect(response.status).toBe(200);
-      expect(mockCreateMeetingToken).toHaveBeenCalledWith(
-        expect.objectContaining({ isOwner: false }),
-      );
+      const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+      expect(res.status).toBe(200);
+      expect(mockCreateMeetingToken).toHaveBeenCalled();
     });
 
-    it("should grant screen share (via isOwner) for admin", async () => {
-      mockAuthenticatedWithProfile("admin-user-id", {
-        role: "admin",
-        first_name: "Admin",
-        username: "admin1",
-      });
-      mockRoomLookup(createSpecialRoom("admin_only"));
+    it("returns 500 when a non-409 Daily error escapes lazy-create", async () => {
+      authAs("admin-id", { role: "admin", first_name: "Boss" });
+      mockTables({ group: {} });
+      mockCreateDailyRoom.mockRejectedValue(new DailyApiError(500, "Daily down"));
 
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-
-      expect(response.status).toBe(200);
-      expect(mockCreateMeetingToken).toHaveBeenCalledWith(
-        expect.objectContaining({ isOwner: true }),
-      );
-    });
-
-    it("should grant screen share (via isOwner) for gedu", async () => {
-      mockAuthenticatedWithProfile("gedu-user-id", {
-        role: "gedu",
-        first_name: "Educator",
-        username: "edu1",
-      });
-      mockRoomLookup(createGroupRoom());
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-
-      expect(response.status).toBe(200);
-      expect(mockCreateMeetingToken).toHaveBeenCalledWith(
-        expect.objectContaining({ isOwner: true }),
-      );
-    });
-
-    it("should deny screen share (via non-owner) for gamer", async () => {
-      mockAuthenticatedWithProfile("gamer-user-id", {
-        role: "gamer",
-        first_name: "Gamer",
-        username: "gamer1",
-      });
-
-      mockAdminFrom.mockImplementation((table: string) => {
-        if (table === "voice_rooms") {
-          return {
-            select: vi.fn().mockReturnValue({
-              eq: vi.fn().mockReturnValue({
-                single: vi.fn().mockResolvedValue(mockSupabaseSuccess(createGroupRoom())),
-              }),
-            }),
-          };
-        }
-        if (table === "group_enrollments") {
-          return {
-            select: vi.fn().mockReturnValue({
-              eq: vi.fn().mockReturnValue({
-                eq: vi.fn().mockReturnValue({
-                  eq: vi.fn().mockReturnValue({
-                    limit: vi.fn().mockReturnValue({
-                      maybeSingle: vi.fn().mockResolvedValue(mockSupabaseSuccess({
-                        id: "enrollment-1",
-                        created_at: new Date(Date.now() - 7 * 24 * 3600_000).toISOString(),
-                      })),
-                    }),
-                  }),
-                }),
-              }),
-            }),
-          };
-        }
-        return {};
-      });
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-
-      expect(response.status).toBe(200);
-      expect(mockCreateMeetingToken).toHaveBeenCalledWith(
-        expect.objectContaining({ isOwner: false }),
-      );
-    });
-
-    it("should set group room token expiry to windowClosesAt + grace period", async () => {
-      mockAuthenticatedWithProfile("gedu-user-id", {
-        role: "gedu",
-        first_name: "Educator",
-        username: "edu1",
-      });
-      mockRoomLookup(createGroupRoom());
-
-      const windowClosesAt = new Date("2026-03-03T15:05:00.000Z");
-      mockComputeSessionWindow.mockReturnValue({
-        isOpen: true,
-        nextSessionStart: new Date("2026-03-03T14:00:00.000Z"),
-        windowOpensAt: new Date("2026-03-03T13:55:00.000Z"),
-        windowClosesAt,
-      });
-
-      await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-
-      const expectedExp = Math.round(windowClosesAt.getTime() / 1000)
-        + VOICE_CONFIG.TOKEN_EXPIRY_GRACE_SECONDS;
-      expect(mockCreateMeetingToken).toHaveBeenCalledWith(
-        expect.objectContaining({ expUnix: expectedExp }),
-      );
-    });
-
-    it("should not set expUnix for always-open rooms", async () => {
-      mockAuthenticatedWithProfile("admin-user-id", {
-        role: "admin",
-        first_name: "Admin",
-        username: "admin1",
-      });
-      mockRoomLookup(createSpecialRoom("admin_only"));
-
-      await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-
-      expect(mockCreateMeetingToken).toHaveBeenCalledWith(
-        expect.objectContaining({ expUnix: undefined }),
-      );
+      const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+      expect(res.status).toBe(500);
     });
   });
 
-  describe("lazy Daily.co room creation", () => {
-    it("should create Daily.co room if it does not exist", async () => {
-      mockAuthenticatedWithProfile("admin-user-id", {
-        role: "admin",
-        first_name: "Admin",
-        username: "admin1",
-      });
-      mockRoomLookup(createSpecialRoom("admin_only"));
-      mockGetDailyRoom.mockResolvedValue(null);
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-
-      expect(response.status).toBe(200);
-      expect(mockGetDailyRoom).toHaveBeenCalledWith("admin-lounge");
-      expect(mockCreateDailyRoom).toHaveBeenCalledWith({ name: "admin-lounge" });
-    });
-
-    it("should skip Daily.co creation if room already exists", async () => {
-      mockAuthenticatedWithProfile("admin-user-id", {
-        role: "admin",
-        first_name: "Admin",
-        username: "admin1",
-      });
-      mockRoomLookup(createSpecialRoom("admin_only"));
-      mockGetDailyRoom.mockResolvedValue({ name: "admin-lounge" });
-
-      const response = await POST(createTokenRequest({ roomId: "room-uuid-1234" }));
-
-      expect(response.status).toBe(200);
-      expect(mockCreateDailyRoom).not.toHaveBeenCalled();
+  describe("response shape", () => {
+    it("returns {token, roomUrl, role}", async () => {
+      authAs("admin-id", { role: "admin", first_name: "Boss" });
+      mockTables({ group: {} });
+      const res = await POST(tokenRequest({ groupId: GROUP_ID }));
+      const data = await res.json();
+      expect(data).toEqual(
+        expect.objectContaining({
+          token: "mock-daily-token",
+          roomUrl: expect.stringContaining("testdomain.daily.co/g-aaaaaaaa-"),
+          role: "admin",
+        }),
+      );
     });
   });
 });
